+++ scripts/lighthouse_score_upload.py
@@ -8,14 +8,11 @@ import sys
 from datetime import datetime
 
 import psycopg2
-import pytest
-from benchmark_helpers import extract_stats_from_json
 
 
 def insert_benchmarking_data(
     db_connection_url: str,
     lighthouse_data: dict,
-    performance_data: list[dict],
     commit_sha: str,
     pr_title: str,
 ):
@@ -24,13 +21,11 @@ def insert_benchmarking_data(
     Args:
         db_connection_url: The URL to connect to the database.
         lighthouse_data: The Lighthouse data to insert.
-        performance_data: The performance data to insert.
         commit_sha: The commit SHA to insert.
         pr_title: The PR title to insert.
     """
     # Serialize the JSON data
     lighthouse_json = json.dumps(lighthouse_data)
-    performance_json = json.dumps(performance_data)
 
     # Get the current timestamp
     current_timestamp = datetime.now()
@@ -38,14 +33,13 @@ def insert_benchmarking_data(
     # Connect to the database and insert the data
     with psycopg2.connect(db_connection_url) as conn, conn.cursor() as cursor:
         insert_query = """
-            INSERT INTO benchmarks (lighthouse, performance, commit_sha, pr_title, time)
-            VALUES (%s, %s, %s, %s, %s);
+            INSERT INTO benchmarks (lighthouse, commit_sha, pr_title, time)
+            VALUES (%s, %s, %s, %s);
             """
         cursor.execute(
             insert_query,
             (
                 lighthouse_json,
-                performance_json,
                 commit_sha,
                 pr_title,
                 current_timestamp,
@@ -91,33 +85,6 @@ def get_lighthouse_scores(directory_path: str) -> dict:
     return scores
 
 
-def run_pytest_and_get_results(test_path=None) -> str:
-    """Runs pytest and returns the results.
-
-    Args:
-        test_path: The path to the tests to run.
-
-    Returns:
-        the name of the output file
-    """
-    # Set the default path to the current directory if no path is provided
-    if not test_path:
-        test_path = os.getcwd()
-
-    output_file = "benchmark_report.json"
-
-    # Ensure you have installed the pytest-json plugin before running this
-    pytest_args = ["-v", f"--benchmark-json={output_file}", test_path]
-
-    # Run pytest with the specified arguments
-    pytest.main(pytest_args)
-
-    # Print ls of the current directory
-    print(os.listdir())
-
-    return output_file
-
-
 def main():
     """Runs the benchmarks and inserts the results into the database."""
     # Get the commit SHA and JSON directory from the command line arguments
@@ -131,17 +98,11 @@ def main():
     if db_url is None or pr_title is None:
         sys.exit("Missing environment variables")
 
-    # Run pytest and get the results
-    result_json_file = run_pytest_and_get_results()
-    cleaned_results = extract_stats_from_json(result_json_file)
-
     # Get the Lighthouse scores
     lighthouse_scores = get_lighthouse_scores(json_dir)
 
     # Insert the data into the database
-    insert_benchmarking_data(
-        db_url, lighthouse_scores, cleaned_results, commit_sha, pr_title
-    )
+    insert_benchmarking_data(db_url, lighthouse_scores, commit_sha, pr_title)
 
 
 if __name__ == "__main__":
